refactor(playground): split websocket client module
