+++ packages/agent-playground/src/lib/playground/websocket-client/index.ts
@@ -1,0 +1,11 @@
+export { PlaygroundWebSocketClient } from './playground-websocket-client';
+export { PLAYGROUND_WS_CLIENT_EVENTS, PLAYGROUND_WS_MESSAGE_TYPES } from './constants';
+export type {
+  IPlaygroundConnectionStatus,
+  IPlaygroundWebSocketAuthenticatedEvent,
+  IPlaygroundWebSocketConnectionEvent,
+  IPlaygroundWebSocketErrorEvent,
+  TPlaygroundWebSocketEventHandler,
+  TPlaygroundWebSocketEventPayload,
+} from './types';
+export type { IPlaygroundWebSocketMessage, TPlaygroundWebSocketMessageKind } from '../types';
